+++ core/dbt/config/profile.py
@@ -91,6 +91,7 @@ class Profile(HasCredentials):
     user_config: UserConfig
     threads: int
     credentials: Credentials
+    profile_env_vars: Dict[str, Any]
 
     def __init__(
         self,
@@ -108,6 +109,7 @@ class Profile(HasCredentials):
         self.user_config = user_config
         self.threads = threads
         self.credentials = credentials
+        self.profile_env_vars = {}  # never available on init
 
     def to_profile_info(
         self, serialize_credentials: bool = False