+++ core/dbt/task/base.py
@@ -80,6 +80,8 @@ class BaseTask(metaclass=ABCMeta):
     @classmethod
     def from_args(cls, args):
         try:
+            # This is usually RuntimeConfig but will be UnsetProfileConfig
+            # for the clean or deps tasks
             config = cls.ConfigType.from_args(args)
         except dbt.exceptions.DbtProjectError as exc:
             logger.error("Encountered an error while reading the project:")