+++ core/dbt/task/deps.py
@@ -4,7 +4,6 @@ import dbt.exceptions
 
 from dbt.config import UnsetProfileConfig
 from dbt.config.renderer import DbtProjectYamlRenderer
-from dbt.context.target import generate_target_context
 from dbt.deps.base import downloads_directory
 from dbt.deps.resolver import resolve_packages
 
@@ -52,9 +51,7 @@ class DepsTask(BaseTask):
         with downloads_directory():
             final_deps = resolve_packages(packages, self.config)
 
-            renderer = DbtProjectYamlRenderer(generate_target_context(
-                self.config, self.config.cli_vars
-            ))
+            renderer = DbtProjectYamlRenderer(self.config, self.config.cli_vars)
 
             packages_to_upgrade = []
             for package in final_deps: