+++ test/integration/068_partial_parsing_tests/test_pp_vars.py
@@ -229,4 +229,108 @@ class EnvVarTest(BasePPTest):
         del os.environ['TEST_SCHEMA_VAR']
         del os.environ['ENV_VAR_COLOR']
         del os.environ['ENV_VAR_SOME_KEY']
+        del os.environ['ENV_VAR_OWNER']
+
+
+class ProjectEnvVarTest(BasePPTest):
+
+    @property
+    def project_config(self):
+        # Need to set the environment variable here initially because
+        # the unittest setup does a load_config.
+        os.environ['ENV_VAR_NAME'] = "Jane Smith"
+        return {
+            'config-version': 2,
+            'seed-paths': ['seeds'],
+            'test-paths': ['tests'],
+            'macro-paths': ['macros'],
+            'seeds': {
+                'quote_columns': False,
+            },
+            'models': {
+                '+meta': {
+                    'meta_name': "{{ env_var('ENV_VAR_NAME') }}"
+                }
+            }
+        }
+
+    @use_profile('postgres')
+    def test_postgres_project_env_vars(self):
+
+        # Initial run
+        self.setup_directories()
+        self.copy_file('test-files/model_one.sql', 'models/model_one.sql')
+        self.run_dbt(['clean'])
+        results = self.run_dbt(["run"])
+        self.assertEqual(len(results), 1)
+        manifest = get_manifest()
+        state_check = manifest.state_check
+        model_id = 'model.test.model_one'
+        model = manifest.nodes[model_id]
+        self.assertEqual(model.config.meta['meta_name'], 'Jane Smith')
+        env_vars_hash_checksum = state_check.project_env_vars_hash.checksum
+
+        # Change the environment variable
+        os.environ['ENV_VAR_NAME'] = "Jane Doe"
+        results = self.run_dbt(["run"])
+        self.assertEqual(len(results), 1)
+        manifest = get_manifest()
+        model = manifest.nodes[model_id]
+        self.assertEqual(model.config.meta['meta_name'], 'Jane Doe')
+        self.assertNotEqual(env_vars_hash_checksum, manifest.state_check.project_env_vars_hash.checksum)
+
+        # cleanup
+        del os.environ['ENV_VAR_NAME']
+
+class ProfileEnvVarTest(BasePPTest):
+
+    @property
+    def profile_config(self):
+        # Need to set these here because the base integration test class
+        # calls 'load_config' before the tests are run.
+        # Note: only the specified profile is rendered, so there's no
+        # point it setting env_vars in non-used profiles.
+        os.environ['ENV_VAR_USER'] = 'root'
+        os.environ['ENV_VAR_PASS'] = 'password'
+        return {
+            'config': {
+                'send_anonymous_usage_stats': False
+            },
+            'test': {
+                'outputs': {
+                    'dev': {
+                        'type': 'postgres',
+                        'threads': 1,
+                        'host': self.database_host,
+                        'port': 5432,
+                        'user': "root",
+                        'pass': "password",
+                        'user': "{{ env_var('ENV_VAR_USER') }}",
+                        'pass': "{{ env_var('ENV_VAR_PASS') }}",
+                        'dbname': 'dbt',
+                        'schema': self.unique_schema()
+                    },
+                },
+                'target': 'dev'
+            }
+        }
+
+    @use_profile('postgres')
+    def test_postgres_profile_env_vars(self):
+
+        # Initial run
+        os.environ['ENV_VAR_USER'] = 'root'
+        os.environ['ENV_VAR_PASS'] = 'password'
+        self.setup_directories()
+        self.copy_file('test-files/model_one.sql', 'models/model_one.sql')
+        results = self.run_dbt(["run"])
+        manifest = get_manifest()
+        env_vars_checksum = manifest.state_check.profile_env_vars_hash.checksum
+
+        # Change env_vars
+        os.environ['ENV_VAR_PASS'] = 'my_pass'
+        (results, log_output) = self.run_dbt_and_capture(["run"], expect_pass=False)
+        self.assertTrue('env vars used in profiles.yml have changed' in log_output)
+        manifest = get_manifest()
+        self.assertNotEqual(env_vars_checksum, manifest.state_check.profile_env_vars_hash.checksum)